+++ server/config/index.js
@@ -1,8 +1,11 @@
 /* eslint no-console:0 */
 /* globals console */
 import path from 'path';
+import moment from 'moment';
 import colors from 'colors/safe';
 import { argv } from 'yargs';
+
+import { normalizePort } from '../utils';
 
 if (argv.mocks) {
   process.env.MOCKS = true;
@@ -16,20 +19,6 @@ if (argv.dd) {
 
 export const isProd = process.env.NODE_ENV === 'production';
 
-const normalizePort = (val) => {
-  const port = parseInt(val, 10);
-
-  if (isNaN(port)) {
-    return val;
-  }
-
-  if (port >= 0) {
-    return port;
-  }
-
-  return false;
-};
-
 export const SERVER_PORT = normalizePort(process.env.PORT || '3000');
 export const WEB_SOCKET_PORT = 4001;
 
@@ -37,3 +26,27 @@ const ROOT = '../';
 export const PUBLIC_PATH = path.join(__dirname, ROOT, 'public');
 export const BUNDLE_PATH = '/dist/bundle.js';
 export const VIEWS_PATH = path.join(__dirname, ROOT, 'views');
+
+/**
+ * Contrib configurations.
+ */
+export const logOptions = {
+  fg: 'green',
+  selectedFg: 'blue',
+  label: 'Server Log',
+  width: '100%',
+  height: '50%',
+  border: { type: 'line', fg: 'red' }
+};
+
+export const tableOptions = {
+  fg: 'white',
+  selectedFg: 'white',
+  selectedBg: 'red',
+  label: `Room statuses as of ${moment().format('LLLL')}`,
+  width: '100%',
+  height: '50%',
+  border: { type: 'line', fg: 'red' },
+  columnSpacing: 10,
+  columnWidth: [20, 40]
+};
