+++ server/utils/index.js
@@ -1,4 +1,5 @@
 export * from '../../universal/utils';
+export * from './server';
 export * from './flash';
 export * from './hardware';
 export * from './logging';
